+++ avidemux/qt4/ADM_userInterfaces/ADM_gui/Q_gui2.cpp
@@ -881,7 +881,7 @@ void UI_setFrameCount(uint32_t curFrame,uint32_t total)
 void UI_updateTimeCount(uint32_t curFrame,uint32_t fps)
 {
 	char text[80];   
-	uint16_t mm,hh,ss,ms;
+	uint32_t mm,hh,ss,ms;
 
 	frame2time(curFrame,fps, &hh, &mm, &ss, &ms);
 	sprintf(text, "%02d:%02d:%02d.%03d", hh, mm, ss, ms);
@@ -895,7 +895,7 @@ void UI_updateTimeCount(uint32_t curFrame,uint32_t fps)
 void UI_setTimeCount(uint32_t curFrame,uint32_t total, uint32_t fps)
 {
 	char text[80];   
-	uint16_t mm,hh,ss,ms;
+	uint32_t mm,hh,ss,ms;
 
 	if(total) total--; // We display from 0 to X
 
@@ -916,7 +916,7 @@ void UI_setTimeCount(uint32_t curFrame,uint32_t total, uint32_t fps)
 void UI_setCurrentTime(uint64_t curTime)
 {
   char text[80];
- uint16_t mm,hh,ss,ms;
+ uint32_t mm,hh,ss,ms;
  uint32_t shorty=(uint32_t)(curTime/1000);
 
     ms2time(shorty,&hh,&mm,&ss,&ms);
@@ -932,7 +932,7 @@ void UI_setCurrentTime(uint64_t curTime)
 void UI_setTotalTime(uint64_t curTime)
 {
   char text[80];
- uint16_t mm,hh,ss,ms;
+ uint32_t mm,hh,ss,ms;
  uint32_t shorty=(uint32_t)(curTime/1000);
 
     ms2time(shorty,&hh,&mm,&ss,&ms);
@@ -946,7 +946,7 @@ void UI_setTotalTime(uint64_t curTime)
 void UI_setMarkers(uint64_t a, uint64_t b)
 {
 	char text[80];
-    uint16_t hh,mm,ss,ms;
+    uint32_t hh,mm,ss,ms;
     uint32_t timems;
     a/=1000;
     b/=1000;
